Crate synchronoise

A collection of synchronization primitives that build on the primitives available in the standard library.

This library contains the following special-purpose synchronization primitives:

CountdownEvent, a primitive that keeps a counter and allows a thread to wait until the counter reaches zero.

SignalEvent, a primitive that allows one or more threads to wait on a signal from another thread.

WriterReaderPhaser, a primitive that allows multiple wait-free “writer critical sections” against a “reader phase flip” that waits for currently-active writers to finish.

Re-exports

pub use event::CountdownEvent;

pub use event::SignalEvent;

pub use phaser::WriterReaderPhaser;

Modules

event

“Event” primitives, allowing one thread to wait on a signal or countdown from other threads.

phaser

Support module for WriterReaderPhaser and related structs.